Delete migration files and addAuthors File.
