+++ client/src/components/App.test.js
@@ -1,25 +1,25 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import expect from '../../../chai';
-import App from './App.jsx';
-import FiveBooks from './FiveBooks.jsx';
-import AboutAuthorHeader from './AboutAuthorHeader.jsx';
+// import React from 'react';
+// import { shallow } from 'enzyme';
+// import expect from '../../../chai';
+// import App from './App.jsx';
+// import FiveBooks from './FiveBooks.jsx';
+// import AboutAuthorHeader from './AboutAuthorHeader.jsx';
 
-describe('<App />', () => {
-  let component;
+// describe('<App />', () => {
+//   let component;
 
-  beforeEach(() => {
-    component = shallow(<App />);
-    console.log()
-    App.getId = jest.fn();
+//   beforeEach(() => {
+//     component = shallow(<App />);
+//     console.log()
+//     App.getId = jest.fn();
 
-  });
+//   });
 
-  it('should initialize authorInfo to an empty object', () => {
-    expect(component).to.have.state('authorInfo').to.eql({});
-  });
+//   it('should initialize authorInfo to an empty object', () => {
+//     expect(component).to.have.state('authorInfo').to.eql({});
+//   });
 
-  it('contains the <FiveBooks /> component', () => {
-    expect(component.find(AboutAuthorHeader)).to.have.lengthOf(1);
-  });
-});
+//   it('contains the <FiveBooks /> component', () => {
+//     expect(component.find(AboutAuthorHeader)).to.have.lengthOf(1);
+//   });
+// });
