+++ client/src/components/FollowButton.jsx
@@ -1,25 +1,26 @@
-import React from 'react';
-import styled from 'styled-components';
+// may remove this componet to place within the 'AuthorName' component
+// import React from 'react';
+// import styled from 'styled-components';
 
-const Button = styled.button`
-  background-color: #f4f1ea;
-  color: #333333;
-  border-radius: 3px;
-  border: 1px solid #D6D0C4;
-  font-family: "Lato", "Helvetica Neue", "Helvetica", sans-serif;
-  font-size: 14px;
-  padding-bottom: 8px;
-  padding-left: 12px;
-  padding-right: 12px;
-  padding-top: 8px;
-`;
+// const Button = styled.button`
+//   background-color: #f4f1ea;
+//   color: #333333;
+//   border-radius: 3px;
+//   border: 1px solid #D6D0C4;
+//   font-family: "Lato", "Helvetica Neue", "Helvetica", sans-serif;
+//   font-size: 14px;
+//   padding-bottom: 8px;
+//   padding-left: 12px;
+//   padding-right: 12px;
+//   padding-top: 8px;
+// `;
 
-class FollowButton extends React.Component {
-  render() {
-    return (
-        <Button>Follow Author</Button>
-    );
-  }
-}
+// class FollowButton extends React.Component {
+//   render() {
+//     return (
+//         <Button>Follow Author</Button>
+//     );
+//   }
+// }
 
-export default FollowButton;
+// export default FollowButton;
